requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
coverage==5.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
djangorestframework==3.14.0
Faker==6.5.2
filelock==3.0.12
graphene==3.3
graphene-django==3.0.0b7
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.0.4
pipenv==2020.11.15
promise==2.3
pycodestyle==2.6.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typing_extensions==4.7.1
validate-email==1.3
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
